MR-129: change in styles and redirection after subscribe
