+++ src/Components/MemberLogged/SubscribeActivities/MemberActivityView/memberActivity.module.css
@@ -1,9 +1,9 @@
-.title{
-  text-align: center;
-  font-size: 2rem;
-  font-weight: 700;
-  color: #373867;
+.containerLoader{
+  display: flex ;
+  flex-direction: column;
+  align-items: center;
 }
+
 .cardContainer{
   display: flex;
   justify-content: center;
