+++ src/Components/MemberLogged/SubscribeActivities/index.jsx
@@ -1,10 +1,14 @@
 import React from 'react';
-import MemberActivityView from '../MemberActivityView';
+import MemberActivityView from './MemberActivityView';
+import styles from './subscribeActivities.module.css';
 
 const SubscribeActivities = () => {
   return (
     <>
-      <MemberActivityView />
+      <div className={styles.container}>
+        <h2>Sign up for a Class</h2>
+        <MemberActivityView />
+      </div>
     </>
   );
 };
