+++ src/Components/MemberLogged/SubscribeActivities/subscribeActivities.module.css
@@ -1,0 +1,10 @@
+.container{
+  display: flex ;
+  flex-direction: column;
+}
+.container h2{
+  text-align: center;
+  font-size: 2rem;
+  font-weight: 700;
+  color: #373867;
+}
